Move tests to ArrowKeysCefTest.ts
